[et-mgmt-tools]  [PATCH] In virt-manager GUI, when the disk is added, VM cannot be started. Hi
When the disk is added to active VM by virt-manager, VM cannot be
restarted after VM is stopped.
--> Bugzilla Bug 239763
    https://bugzilla.redhat.com/bugzilla/show_bug.cgi?id=239763

This patch fixes it.

When the device is attached to active VM, virt-manager adds the device
twice in add_device() of "domain.py".
1. Call attachDevice()
2. Call define_domain() after the XML is updated
Therefore, the same device is doubly attached.
(So, when the device is disk, VM cannot be started)

This patch fixes "domain.py" as follows so that the same device is not
doubly attached.
  VM is active   : Call attachDevice() only
  VM is inactive : Call define_domain() after the XML is updated


Signed-off-by: Masayuki Sunou <fj1826dm@aa.jp.fujitsu.com>

Thanks,
Masayuki Sunou.

----------------------------------------------------------------------