I thought that the following is the problem.
----------------------------------------------------------------------
In finish() of "create.py", the port number of VNC become -1.
When "Status monitoring" is short, the port number of VNC is updated
in a short time, but, when "Status monitoring" is long, the port number
of VNC is not updated for a long time.
Because the XML data cached in "domain.py" is updated only in tick().
----------------------------------------------------------------------

And, I forgot retrying in "console.py".
Therefore, I adopted updating the port number of VNC in "cerate.py".

I remake the patch.
This patch only clears XML in get_graphics_console(), and the connection
to the console is not retried in "create.py".
So, by applying this patch, the XML data is cleared by retrying in "console.py"
even when "Status monitoring" is long, and the port number of VNC is updated.
(Because retrying in "console.py" calls get_graphics_console())

Thanks,
Masayuki Sunou.

----------------------------------------------------------------------