Separate local net device polling from connection initialization.
Currently we connect to dbus and poke sysfs to find local net devices and
bridges: unfortunately we do this everytime _any_ connection object is
initialized. This floods the debug logs, and is complete overkill, since
net devices are associated with the host and not a connection.

Run the initialization once (at app startup) and carry around only one
instance of the polling code.