manager: Remove Memory, ID, and VCPU rows
VCPU count is pointless to have in the main view. ID is similarly pointless,
its only value being that it showed the connection driver, which is now
part of the connection name.

Memory is the only useful field here, but even it's prescence is dubious:
there isn't much value in seeing at a glance what memory amounts VMs
are statically defined with.