details: Break out VNC + console pages to a separate class.
The details class is getting pretty big, so split out the VNC functionality
into a separate class (like it was many moons ago).

The split isn't as clean as it should be, but things work and readability is
greatly improved. If we ever add support for a different viewer widget (rdp
or spice?) we would need to do something similar anyways.