Add code for virt-manager TUI from OVirt node project
This code is taken from the ovirt node project. Overtime the plan is to share
lots of code between the GUI and TUI. This is step 1 though.

Code can be invoked with:

cd src/virtManagerTui
python nodeadmin.py

Additional requirements beyond plain virt-manager are a system installed IPy
and newt-python (or whatever package provides the 'snack' module)