prefs: Remove console popup option
This was originally requested when the console window didn't have option to
start/stop a VM (see https://bugzilla.redhat.com/show_bug.cgi?id=211385).
That hasn't been valid for a long time. The option was pretty unclear in
the prefs dialog anyways, and had weird behavior like 'open the domain for
all running vms when we start virt-manager'

Just always use the current default of opening the console for VMs that
we create with New VM.