manager: Option to view VM CPU usage of guest OR host
Since forever, all we've ever shown is a VMs CPU usage according to host
resources. However I don't think this is the intuitive default, when users
see a graph for each VM IMO it's implies the CPU usage of the guest's
resources.

Change the default to 'guest CPU usage', and add an extra optional column
for explicitly listing 'host CPU usage'