Add inspection data to the domain object.
Create a new (sub-)object vmmInspectionData, and add an instance of
this data to each domain object.  This data is passed up from from the
inspection thread.

Also an inspection-changed signal has been added so that UI elements
can find out when inspection has finished for a particular VM and
update the UI.