details: Confirm with user before throwing away changes on page change
We've had many users be confused by changes in the details pages not
taking immediate effect, or just forgetting to hit apply.

Now by default we prompt users if switching pages if there are unapplied
changes. This can be disabled like other prompts in Edit->Prefs or with
a 'Don't warn me again' chkbox in the dialog.