TUI: Drop direct HAL usage, use virt-manager code
virt-manager already has code for listing CDROMs and host network devices
via libvirt's nodedev APIs (and optionally falling back to HAL if those
aren't available).

Use this in the TUI so we can drop the hard requirement on HAL, which is
deprecated.

side note: on f14, the TUI segfaults if no cdrom media devs are found when
trying to install a new guest. this issue predates this patch and is likely
a problem in newt or the python_newt bindings. I didn't really dig into
it though.