virt-manager: Show if the shut off vm has a saved image
Hi,

So far in virt-manager one had to click on a shut off vm in the
"Shutoff" state to see if it has a saved image: once the item is
selected then the "play" icon will be "Run" or "Restore". This means
that in case one has a lot of VMs, it's hard to get an overview of what
VM has a saved image and what not.

The attached patch changes the "Shutoff" status message to "Saved" in
case the vm has a saved image.

Thanks.

>From 14445dc510fcc55c267649295246984972691885 Mon Sep 17 00:00:00 2001
From: Miklos Vajna <vmiklos@frugalware.org>
Date: Sat, 30 Jul 2011 22:02:48 +0200
Subject: [PATCH] domain: Show if the shut off vm has a saved image