virt-manager: Throw appropriate error when fs target string already exists.
In case of qemu for default fs type (mount) target is not
actually a directory, it is merely a arbitrary string tag
that is exported to the guest as a hint for where to mount.

This tag is unique and is exported as a virito-9p-pci
device and hence this patch adds check to ensure the target
string/tag is never entered duplicate by the user when
adding 2 or more filesystem devices.

Signed-off-by: Deepak C Shetty <deepakcs@linux.vnet.ibm.com>