console: Fix hang when reconnecting to remote VNC console
VNC disconnect signals aren't being delivered, meaning we never cleanup
up SSH tunnels for remote VNC connections. This causes a reconnect
attempt to hang, trying to reuse the existing tunnels (which doesn't
work).

The problem is that we are cleaning up the VNC widget too early,
unregistering callbacks before they get a chance to be invoked.

Not cleaning up only causes a small memory leak for every console we
open, so only a minor problem compared to the original issue. A more
involved fix is required to clean up safely.