Top level windows shouldn't be visible by default
Causes first run of dialogs to fail with a flicker on F16 KDE at least.

Choosecd also was using 'show' instead of 'present', which combined
with the first run fail meant that the dialog couldn't be displayed on
KDE :(

Also, the progress dialog was unconditionally shown, so this should fix
random flicker for quick operations, even in gnome.