Drop use of gtk threading
In general it complicates things for minor simplifications in some
cases. I think it's better just to learn that anything invoked from
a thread that _might_ touch the UI should just be dispatched in an
idle callback. Anything UI bits that we need immediately in a thread
can just use some creative trickery (like is done in connectauth) to
block in a safe manner.