add MANIFEST.in to specify includes and excludes
With no MANIFEST.in, distutils/setup/sdist gets the filelist
mostly correct except that it adds in whatever is in the
top directory.  Using git-ls-files is not a good solution
either because a number of additional files (not git
managed) will be needed.

The solution is to use a MANIFEST.in.  All of the
toplevel files which are to be included will need to be
specified.  Also, each directory will need to be specified
and unwanted files (e.g., *.pyc) excluded.  Other (non-specified)
files and directories will be ignored.
.
Signed-off-by: Gene Czarcinski <gene@czarc.net>