Add controller example into virt-install manual page
There was an example for USB controller, but to show all the
sub-options, there was the master= option used without extra
information about the fact that it is used for companion controllers
only.  So this patch adds second example in front of the previous one
and re-words the description a little bit.