host: unselect_all members before clear model
Interface_selected is being triggered by model.clear() in
populate_interfaces. It usually calls interface_selected
on every interface after the current selected row.

So the gtk code is probably removing the first
row over and over, and when it removes the
current selection (the interface we are deleting),
it selects the next row in line.
It will raise a KeyError.

This patch will fix this issue.

Signed-off-by: Chen Hanxiao <chenhanxiao@cn.fujitsu.com>