ui: add a warning message for hyper-threading support
Not all CPUs support Hyper-Threading. This patch adds a warning
message next to the threads selection field of processor topology
box. It warns that "Selected CPU model does not support
Hyper-Threading" when # of Threads > 1 and the selected CPU model
does not support Hyper-Threading.

This UI modification was added using glade 3.18.3.

Signed-off-by: Wei Huang <wei@redhat.com>