domain: migrate: Use UNDEFINE_SOURCE
Leaving a VM defined on the source is dangerous: if they are sharing
storage, and aren't using virtlockd, then they might be inadvertently
started in two places at once and trash a disk image.

Some people might still need the option to turn this off, so maybe we
will revive it. Not sure though