addstorage: remove _check_ideal_path
This feature has been added few years ago.  I don't think, that it's a
good feature, as it can ask a user to use different storage than he
actually wants to use.  One thing is automatically create a new storage
for user, if he let as do that, but we shouldn't annoy a user with this
question as he probably don't want to use the proposed storage.  For
example he would like to use different storage pool or while importing
existing storage.

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1232599

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>